bspwm forever

+++ install.sh
@@ -16,17 +16,11 @@ for dir in */; do
   echo -e "${GREEN}Linked ${dir%/}!${END}"
 done
 
-for dot in .bashrc .xinitrc; do
+for dot in .*; do
+  [[ $dot == .git ]] && continue
   echo -e "${BLUE}Removing ${HOME}/${dot}${END}"
   rm -rf "${HOME}/${dot}"
   echo -e "${BLUE}Creating symlink ${HOME}/${dot} -> $(pwd)/${dot}${END}"
   ln -s "$(pwd)/${dot}" "${HOME}/${dot}"
   echo -e "${GREEN}Linked ${dot}!${END}"
 done
-
-echo -e "${BLUE}Removing ${HOME}/.icons${END}"
-rm -rf "${HOME}/.icons"
-echo -e "${BLUE}Creating symlink ${HOME}/.icons/default/index.theme -> $(pwd)/index.theme${END}"
-mkdir -p "${HOME}/.icons/default"
-ln -s "$(pwd)/index.theme" "${HOME}/.icons/default/index.theme"
-echo -e "${GREEN}Linked index.theme!${END}"
